refactor: extract DeFi category badge config to shared utility and apply category colors
Extract duplicated CATEGORY_CONFIG, DEFAULT_CATEGORY_CONFIG, and getCategoryConfig()
from Protocol.tsx and DeFiProtocolDetails.tsx into a shared utility at
kit/src/utils/defiCategoryConfig.ts. Apply per-category color/emoji badges
in both list and detail views.
